// src/Store/User/User.types.ts
// Reducer type
enum USER_ACTION_TYPE {
    SET_USER = " SET_USER",
    CHECK_USER_SESSION = "CHECK_USER_SESSION",
    GOOGLE_LOG_IN_START = "GOOGLE_LOG_IN",
    EMAIL_LOG_IN_START = "EMAIL_LOG_IN_START",
    LOG_IN_SUCCESS = "LOG_IN_SUCCESS",
    LOG_IN_FAILED = "LOG_IN_FAILED",
    SIGN_UP_START = "SIGN_UP_START",
    SIGN_UP_SUCCESS = "SIGN_UP_SUCCESS",
    SIGN_UP_FAILED = "SIGN_UP_FAILED",
    LOG_OUT_START =  "LOG_OUT_START",
    LOG_OUT_SUCCESS =  "LOG_OUT_SUCCESS",
    LOG_OUT_FAILED = "LOG_OUT_FAILED",
};
  
export default USER_ACTION_TYPE;